update getBalance method imports
