+++ packages/substrate/src/__test__/utils/getNativeTokenBalance.test.ts
@@ -2,7 +2,7 @@ import type { ApiPromise } from '@polkadot/api';
 import type { AccountInfo } from '@polkadot/types/interfaces';
 import { BN } from '@polkadot/util';
 
-import { getNativeTokenBalance } from '../substrate/balances.js';
+import { getNativeTokenBalance } from '../../utils/getNativeTokenBalance.js';
 
 describe('getNativeTokenBalance', () => {
   let mockApi: ApiPromise;
