+++ packages/substrate/src/utils/index.ts
@@ -3,4 +3,5 @@ export * from './getPercentageFee.js';
 export * from './getBasicFee.js';
 export * from './deposit/deposit.js';
 export * from './getAssetBalance.js';
+export * from './getNativeTokenBalance.js';
 export * from './deposit/handleTxExtrinsicResult.js';
